move function to utils.py
