+++ sdv/utils/__init__.py
@@ -1,6 +1,6 @@
 """Utils module."""
 
-from sdv.utils.poc import drop_unknown_references
+from sdv.utils.utils import drop_unknown_references
 
 __all__ = (
     'drop_unknown_references',
