+++ sdv/utils/poc.py
@@ -1,72 +1,22 @@
-"""Utility functions."""
-import sys
-from copy import deepcopy
+"""POC functions to use HMASynthesizer succesfully."""
+import warnings
 
-import pandas as pd
-
-from sdv._utils import _validate_foreign_keys_not_null
-from sdv.errors import InvalidDataError, SynthesizerInputError
+from sdv.errors import InvalidDataError
 from sdv.metadata.errors import InvalidMetadataError
 from sdv.multi_table.hma import MAX_NUMBER_OF_COLUMNS
 from sdv.multi_table.utils import (
-    _drop_rows, _get_total_estimated_columns, _print_simplified_schema_summary,
-    _print_subsample_summary, _simplify_data, _simplify_metadata, _subsample_data)
+    _get_total_estimated_columns, _print_simplified_schema_summary, _print_subsample_summary,
+    _simplify_data, _simplify_metadata, _subsample_data)
+from sdv.utils.utils import drop_unknown_references as utils_drop_unknown_references
 
 
-def drop_unknown_references(data, metadata, drop_missing_values=True, verbose=True):
-    """Drop rows with unknown foreign keys.
-
-    Args:
-        data (dict):
-            Dictionary that maps each table name (string) to the data for that
-            table (pandas.DataFrame).
-        metadata (MultiTableMetadata):
-            Metadata of the datasets.
-        drop_missing_values (bool):
-            Boolean describing whether or not to also drop foreign keys with missing values
-            If True, drop rows with missing values in the foreign keys.
-            Defaults to True.
-        verbose (bool):
-            If True, print information about the rows that are dropped.
-            Defaults to True.
-
-    Returns:
-        dict:
-            Dictionary with the dataframes ensuring referential integrity.
-    """
-    success_message = 'Success! All foreign keys have referential integrity.'
-    table_names = sorted(metadata.tables)
-    summary_table = pd.DataFrame({
-        'Table Name': table_names,
-        '# Rows (Original)': [len(data[table]) for table in table_names],
-        '# Invalid Rows': [0] * len(table_names),
-        '# Rows (New)': [len(data[table]) for table in table_names]
-    })
-    metadata.validate()
-    try:
-        metadata.validate_data(data)
-        if drop_missing_values:
-            _validate_foreign_keys_not_null(metadata, data)
-
-        if verbose:
-            sys.stdout.write(
-                '\n'.join([success_message, '', summary_table.to_string(index=False)])
-            )
-
-        return data
-    except (InvalidDataError, SynthesizerInputError):
-        result = deepcopy(data)
-        _drop_rows(result, metadata, drop_missing_values)
-        if verbose:
-            summary_table['# Invalid Rows'] = [
-                len(data[table]) - len(result[table]) for table in table_names
-            ]
-            summary_table['# Rows (New)'] = [len(result[table]) for table in table_names]
-            sys.stdout.write('\n'.join([
-                success_message, '', summary_table.to_string(index=False)
-            ]))
-
-        return result
+def drop_unknown_references(data, metadata):
+    """Wrap the drop_unknown_references function from the utils module."""
+    warnings.warn(
+        "Please access the 'drop_unknown_references' function directly from the sdv.utils module"
+        'instead of sdv.utils.poc.', FutureWarning
+    )
+    return utils_drop_unknown_references(data, metadata)
 
 
 def simplify_schema(data, metadata, verbose=True):
