+++ tests/integration/utils/test_poc.py
@@ -7,11 +7,10 @@ import pandas as pd
 import pytest
 
 from sdv.datasets.demo import download_demo
-from sdv.errors import InvalidDataError
 from sdv.metadata import MultiTableMetadata
 from sdv.multi_table.hma import MAX_NUMBER_OF_COLUMNS, HMASynthesizer
 from sdv.multi_table.utils import _get_total_estimated_columns
-from sdv.utils.poc import drop_unknown_references, get_random_subset, simplify_schema
+from sdv.utils.poc import get_random_subset, simplify_schema
 
 
 @pytest.fixture
@@ -63,98 +62,6 @@ def data():
         'parent': parent,
         'child': child
     }
-
-
-def test_drop_unknown_references(metadata, data, capsys):
-    """Test ``drop_unknown_references`` end to end."""
-    # Run
-    expected_message = re.escape(
-        'The provided data does not match the metadata:\n'
-        'Relationships:\n'
-        "Error: foreign key column 'parent_id' contains unknown references: (5)"
-        ". Please use the utility method 'drop_unknown_references' to clean the data."
-    )
-    with pytest.raises(InvalidDataError, match=expected_message):
-        metadata.validate_data(data)
-
-    cleaned_data = drop_unknown_references(data, metadata)
-    metadata.validate_data(cleaned_data)
-    captured = capsys.readouterr()
-
-    # Assert
-    pd.testing.assert_frame_equal(cleaned_data['parent'], data['parent'])
-    pd.testing.assert_frame_equal(cleaned_data['child'], data['child'].iloc[:4])
-    assert len(cleaned_data['child']) == 4
-    expected_output = (
-        'Success! All foreign keys have referential integrity.\n\n'
-        'Table Name  # Rows (Original)  # Invalid Rows  # Rows (New)\n'
-        '     child                  5               1             4\n'
-        '    parent                  5               0             5'
-    )
-    assert captured.out.strip() == expected_output
-
-
-def test_drop_unknown_references_valid_data(metadata, data, capsys):
-    """Test ``drop_unknown_references`` when data has referential integrity."""
-    # Setup
-    data = deepcopy(data)
-    data['child'].loc[4, 'parent_id'] = 2
-
-    # Run
-    result = drop_unknown_references(data, metadata)
-    captured = capsys.readouterr()
-
-    # Assert
-    pd.testing.assert_frame_equal(result['parent'], data['parent'])
-    pd.testing.assert_frame_equal(result['child'], data['child'])
-    expected_message = (
-        'Success! All foreign keys have referential integrity.\n\n'
-        'Table Name  # Rows (Original)  # Invalid Rows  # Rows (New)\n'
-        '     child                  5               0             5\n'
-        '    parent                  5               0             5'
-    )
-    assert captured.out.strip() == expected_message
-
-
-def test_drop_unknown_references_drop_missing_values(metadata, data, capsys):
-    """Test ``drop_unknown_references`` when there is missing values in the foreign keys."""
-    # Setup
-    data = deepcopy(data)
-    data['child'].loc[4, 'parent_id'] = np.nan
-
-    # Run
-    cleaned_data = drop_unknown_references(data, metadata)
-    metadata.validate_data(cleaned_data)
-    captured = capsys.readouterr()
-
-    # Assert
-    pd.testing.assert_frame_equal(cleaned_data['parent'], data['parent'])
-    pd.testing.assert_frame_equal(cleaned_data['child'], data['child'].iloc[:4])
-    assert len(cleaned_data['child']) == 4
-    expected_output = (
-        'Success! All foreign keys have referential integrity.\n\n'
-        'Table Name  # Rows (Original)  # Invalid Rows  # Rows (New)\n'
-        '     child                  5               1             4\n'
-        '    parent                  5               0             5'
-    )
-    assert captured.out.strip() == expected_output
-
-
-def test_drop_unknown_references_not_drop_missing_values(metadata, data):
-    """Test ``drop_unknown_references`` when the missing values in the foreign keys are kept."""
-    # Setup
-    data['child'].loc[3, 'parent_id'] = np.nan
-
-    # Run
-    cleaned_data = drop_unknown_references(
-        data, metadata, drop_missing_values=False, verbose=False
-    )
-
-    # Assert
-    pd.testing.assert_frame_equal(cleaned_data['parent'], data['parent'])
-    pd.testing.assert_frame_equal(cleaned_data['child'], data['child'].iloc[:4])
-    assert pd.isna(cleaned_data['child']['parent_id']).any()
-    assert len(cleaned_data['child']) == 4
 
 
 def test_simplify_schema(capsys):
@@ -337,7 +244,7 @@ def test_get_random_subset_with_missing_values(metadata, data):
     data['child'].loc[4, 'parent_id'] = np.nan
     expected_warning = re.escape(
         'The data contains null values in foreign key columns. '
-        'We recommend using ``drop_unknown_foreign_keys`` method from sdv.utils.poc'
+        'We recommend using ``drop_unknown_foreign_keys`` method from sdv.utils'
         ' to drop these rows before using ``get_random_subset``.'
     )
